# Switch painless dynamic calls to invokedynamic, remove perf hack/cheat