… modules/lang-painless/src/main/java/org/elasticsearch/painless/PainlessError.java
… modules/lang-painless/src/main/java/org/elasticsearch/painless/PainlessScriptEngineService.java
… modules/lang-painless/src/main/java/org/elasticsearch/painless/WriterConstants.java
… modules/lang-painless/src/main/java/org/elasticsearch/painless/WriterExternal.java
… modules/lang-painless/src/main/java/org/elasticsearch/painless/WriterStatement.java
… modules/lang-painless/src/test/java/org/elasticsearch/painless/BasicAPITests.java
… modules/lang-painless/src/test/java/org/elasticsearch/painless/BasicStatementTests.java
… modules/lang-painless/src/test/java/org/elasticsearch/painless/DynamicCallSiteTests.java
… modules/lang-painless/src/test/java/org/elasticsearch/painless/FieldTests.java
… modules/lang-painless/src/test/java/org/elasticsearch/painless/NoSemiColonTests.java
… modules/lang-painless/src/test/java/org/elasticsearch/painless/WhenThingsGoWrongTests.java
… modules/lang-painless/src/test/resources/rest-api-spec/test/plan_a/15_update.yaml
… modules/lang-painless/src/test/resources/rest-api-spec/test/plan_a/16_update2.yaml
… modules/lang-painless/src/test/resources/rest-api-spec/test/plan_a/25_script_upsert.yaml
… modules/lang-painless/src/test/resources/rest-api-spec/test/plan_a/30_search.yaml