Currently using any dynamic language features of painless is slow. We even have in docs recommending people add a lot of casts to work around this. The hack in https://github.com/elastic/elasticsearch/pull/18169 speeds up access to a document's fields, but this is fragile and not a general solution.

This PR switches all dynamic calls, dynamic loads, and dynamic stores over to use invokedynamic with inline caching (cascaded, goes megamorphic at 5 types). I based it on this one (BSD license) https://github.com/qmx/jsr292-cookbook/blob/master/inlining-cache/src/jsr292/cookbook/icache/RT.java

We don't need to do anything with dynamic array load/store, its already a very simple 3-way decision tree and never involves a dynamic lookup. Same goes for the operators: i left that stuff as is.

I removed the instanceof-performance hack for accessing a document's fields, we don't need it anymore.

I cleaned up some getter/setter stuff too, we compute these at compile time. Also its now possible to get access to `isEmpty()`-style methods as `.empty`, we need that to support the scripting api the way its documented for groovy at least. 

For the nightly benchmark script I see:
SEARCH expression (median): 0.271813 sec
SEARCH painless_static (median): 0.517333 sec
SEARCH painless_dynamic (median): 0.582639 sec

So we should clean up and simplify the documentation, I don't think the dynamic access is trappy anymore, and we shouldn't recommend people add a bunch of casts to their scripts to try to make them faster. If someone wants to eek out extra speed, instead they should use lucene expressions which bypasses the slow scripting API.