@rmuir 2016-05-09
I cleaned up and ran benchmarks: this is currently a script that runs on all of geonames and looks like this: `(Math.log(Math.abs(input.doc['population'].value)) + input.doc['elevation'].value * input.doc['latitude'].value)/_score`

I altered the way we access document fields and values to show perf for different dynamic language features:

| style | median time (master) | median time (branch) |
| --- | --- | --- |
|`doc['field'].value`|0.631238|0.582639|
|`doc['field'].getValue()`|4.148362|0.609807|
|`doc.field.0`|5.567170|0.603587|
|`doc.field.get(0)`|6.588988|0.606248|

`.value` shows the hack from #18169 working: we remove the hack here and have the same performance. 
`.getValue()`: no-arg dynamic method
`.field.0`: two field loads, one that was especially slow before (the .0)
`.field.get(0)`: field load + dynamic method with parameter

The static version (with casts) is ~ `0.517333 s`, so there isn't much overhead. I benchmark with the scripts `inline` so these times include compilation and JIT costs.