@rmuir 2016-05-09
Also I want to defer trying to "improve" the signature we create here. Currently we often know the real types of parameters, e.g. `get(0)`, we know the one parameter is an int, we don't need to box it, but we explicitly do a lot of conversions anyway. That is because the current code is setup to work that way. It is silly to keep it at Object long-term but I want to fix the super slow calls first. 

We can later refactor to improve that (e.g. all tests pass if i do this: https://github.com/rmuir/elasticsearch/commit/e029b9ec546f872b7ee093a9ef8d6180b9709318 but thats really hacky).