@mikemccand 2016-05-09
We are now running simple script search performance tests in the nightly classic benchmarks: https://benchmarks.elastic.co/index.html#search_qps_scripts

It shows already a big speedup (I will add annots!) for painless dynamic from the last few improvements.