@uschindler 2016-05-09
Hi Robert,
I created the MethodHandle code for the array.length bytecode: https://gist.github.com/uschindler/916930741f245204e418c56025a8b58a

This code has a static method where you can pass an primitive or x-extends-Object array and you get a MethodHandle to retun its length. This should be faster than the reflective Array.getLength (because it is statically compiled and we have a separate method for every primitive type). Maybe we shold benchmark it.

Basically it holds a static map array->type to methodhandle (dynamically build) for primitives and Object[], and if the actual type is not found there (e.g, String[]), it adds a necessary cast to Object[].

The code has test as main() method inline.