@rmuir 2016-05-09
Thanks Uwe, I will fold in your ArrayUtil class instead of calling the slow Array.getLength.