@uschindler 2016-05-09
+1 as a first step. LGTM